+++ models/ProfessionalSchool.cs
@@ -5,7 +5,7 @@ namespace NotasApi.models
     public class ProfessionalSchool
     {
         [Key]
-        public int IdProfSchool { get; set; }
+        public long IdProfSchool { get; set; }
         public string Name { get; set; }
 
         public ICollection<Career>? Careers {get;set;}
