+++ models/Semester.cs
@@ -5,7 +5,7 @@ namespace NotasApi.models
     public class Semester
     {
         [Key]
-        public int IdSemester { get; set; }
+        public long IdSemester { get; set; }
         public string NameSemester { get; set; }
 
         public ICollection<TeacherCourse>? TeacherCourses {get;set;}
